dw_batch/pyproject.toml
[tool.uv]
link-mode = "copy"  # Avoid hardlink warnings on Windows/OneDrive

[project]
name = "batch-summary-doubleword"
version = "0.1.0"
description = "Batch document processing using Doubleword API with support for PDF, Excel, CSV, DOCX, PPTX, and more"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pypdf>=6.6.0",
    "pdfplumber>=0.11.9",
    "openai>=2.14.0",
    "python-dotenv>=1.1.0",
    "python-docx>=1.1.0",
    "python-pptx>=1.0.0",
    "odfpy>=1.4.1",
    "pandas>=2.0.0",
    "xlrd>=2.0.0",
    "pdf2image>=1.16.0",
    "Pillow>=10.0.0",
]
